requirements.txt
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
isort==5.12.0
numpy==1.24.2
pandas==2.0.0
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
six==1.16.0
tomli==2.0.1
tzdata==2023.3
urllib3==1.26.15
